refactor(ui): arrangement clips tranche 1: selection and clip operations
Tranche 3b part 1: thirteen more Message variants join the
arrangement domain: clip selection (single/multi via shift),
remove, audio/note clip timeline moves, resize, cross-track move,
audio clip loop toggle/region, time selection, selection-as-loop,
and delete/duplicate of the multi-selection. Time-selection state
and the drag auto-scroll flag join the arrangement slice.

Mechanics of note for the guide:
- Arms were extracted and transformed by script (brace-balanced,
  '=>'-guarded) with hand passes only where flagged (transport
  coupling, App-level scrolling).
- New action fields carry the cross-domain effects: focus the
  clip tab, auto-scroll to a dragged clip, promote the time
  selection to the transport loop (the router replays it through
  the transport domain rather than mutating transport state
  directly).
- ArrangementCtx supplies samples_per_beat for drag snapping.

Suite green at 21/21 with the hardened gate (a compile-failing
test target was caught twice by requiring the ok-count).
